handle failed cases

+++ eval.py
@@ -23,18 +23,33 @@ def main (hyp_file, ref_file):
 				s1 = tokens1[-2].strip()
 				s2 = tokens2[-2].strip()
 
-				with codecs.open('hyp', mode='w', encoding='utf-8') as of: of.write(s1)
-				with codecs.open('ref', mode='w', encoding='utf-8') as of: of.write(s2)
+				c1 = tokens1[-1].strip()
+				c2 = tokens2[-1].strip()
 
-				cmd = 'python WERpp/wer++.py --cer {} {}'.format('hyp', 'ref')
-				ret = subprocess.Popen(cmd.split(' '), stdout=subprocess.PIPE).communicate()[0]
-				cer = ret.split(' ')[1]
+				if float(c1) == 0 or float(c2) == 0:
+					if   float(c1) == 0 and float(c2) == 0:
+						ret = 'Failed: hyp + ref'
+						cer = -3
+					elif float(c2) == 0:
+						ret = 'Failed: ref'
+						cer = -2
+					else:
+						ret = 'Failed: hyp'
+						cer = -1
+				else:
+					with codecs.open('hyp', mode='w', encoding='utf-8') as of:
+						of.write(s1 + '\n')
+					with codecs.open('ref', mode='w', encoding='utf-8') as of:
+						of.write(s2 + '\n')
+
+					cmd = 'python WERpp/wer++.py --cer {} {}'.format('hyp', 'ref')
+					ret = subprocess.Popen(cmd.split(' '), stdout=subprocess.PIPE).communicate()[0].strip()
+					cer = ret.split(' ')[1]
 
 				print '{}, {}, {}, {}'.format(t1, t2, cer, ret)
-				break
 
-			os.remove('ref')
-			os.remove('hyp')
+			if os.path.exists('ref'): os.remove('ref')
+			if os.path.exists('hyp'): os.remove('hyp')
 
 if __name__ == '__main__':
 	parser = argparse.ArgumentParser()
